+++ strato_query/__init__.py
@@ -8,3 +8,8 @@ Notes :
 
 December 26, 2018
 """
+
+from .api_query import SDAPIQuery
+from .authentication import authenticate_to_api
+from .filters import *
+from .query_structures import *